Core/Spells: Extended spell required shapeshift masks to 64 bits and defined a new spell attribute

+++ src/server/game/Spells/SpellInfo.h
@@ -345,8 +345,8 @@ public:
     uint32 AttributesEx12;
     uint32 AttributesEx13;
     uint32 AttributesCu;
-    uint32 Stances;
-    uint32 StancesNot;
+    uint64 Stances;
+    uint64 StancesNot;
     uint32 Targets;
     uint32 TargetCreatureType;
     uint32 RequiresSpellFocus;
